+++ demo/DemoV2/Model/Base/Flags.generated.cs
@@ -8,7 +8,7 @@ using Newtonsoft.Json.Linq;
 
 namespace DemoV2.Model.Base {
 /// AUTOGENERED BY caffoa ///
-    public partial class Flags {
+    public partial class Flags : IEquatable<Flags> {
         public const string FlagsObjectName = "flags";
         [JsonProperty("id")]
         public virtual string Id { get; set; }
@@ -22,5 +22,19 @@ namespace DemoV2.Model.Base {
             Desc = other.Desc;
         }
         public Flags ToFlags() => new Flags(this);
+        public bool Equals(Flags other) {
+            if (ReferenceEquals(null, other)) return false;
+            if (ReferenceEquals(this, other)) return true;
+            return Id == other.Id && Desc == other.Desc;
+        }
+        public override bool Equals(object obj) => Equals(obj as Flags);
+        public override int GetHashCode() {
+            var hashCode = new HashCode();
+            hashCode.Add(Id);
+            hashCode.Add(Desc);
+            return hashCode.ToHashCode();
+        }
+        public static bool operator==(Flags a, Flags b) => Equals(a, b);
+        public static bool operator!=(Flags a, Flags b) => !Equals(a, b);
     }
 }