+++ demo/DemoV2/Model/EnumObject.generated.cs
@@ -11,7 +11,7 @@ using DemoV2.Model.Base;
 
 namespace DemoV2.Model {
 /// AUTOGENERED BY caffoa ///
-    public partial class EnumObject {
+    public partial class EnumObject : IEquatable<EnumObject> {
         public const string EnumObjectObjectName = "enumObject";
         [JsonProperty("single")]
         public virtual MyEnumType Single { get; set; }
@@ -41,5 +41,20 @@ namespace DemoV2.Model {
             AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, object>(other.AdditionalProperties) : null;
         }
         public EnumObject ToEnumObject() => new EnumObject(this);
+        public bool Equals(EnumObject other) {
+            if (ReferenceEquals(null, other)) return false;
+            if (ReferenceEquals(this, other)) return true;
+            return Single == other.Single && WithDefault == other.WithDefault && Array.SequenceEqual(other.Array) && Nullable == other.Nullable && NullableReferenced == other.NullableReferenced;
+        }
+        public override bool Equals(object obj) => Equals(obj as EnumObject);
+        public override int GetHashCode() {
+            var hashCode = new HashCode();
+            hashCode.Add((int) Single);
+            hashCode.Add((int) WithDefault);
+            hashCode.Add(Array);
+            hashCode.Add((int) Nullable);
+            hashCode.Add((int) NullableReferenced);
+            return hashCode.ToHashCode();
+        }
     }
 }