+++ demo/DemoV2/Model/ExtendedAddress.generated.cs
@@ -12,7 +12,7 @@ using DemoV2.Model.Base;
 
 namespace DemoV2.Model {
 /// AUTOGENERED BY caffoa ///
-    public partial class ExtendedAddress {
+    public partial class ExtendedAddress : IEquatable<ExtendedAddress> {
         public const string ExtendedAddressObjectName = "extendedAddress";
         [JsonProperty("street", Required = Required.Always)]
         public virtual string Street { get; set; }
@@ -59,5 +59,23 @@ namespace DemoV2.Model {
             Flags = deepClone ? other.Flags?.ToDictionary(entry => entry.Key, entry => entry.Value?.ToFlags()) : other.Flags;
         }
         public ExtendedAddress ToExtendedAddress() => new ExtendedAddress(this);
+        public bool Equals(ExtendedAddress other) {
+            if (ReferenceEquals(null, other)) return false;
+            if (ReferenceEquals(this, other)) return true;
+            return Street == other.Street && StreetExtra == other.StreetExtra && PostalCode == other.PostalCode && City == other.City && Country == other.Country && AddressType == other.AddressType && Flags.SequenceEqual(other.Flags) && AddressType2 == other.AddressType2;
+        }
+        public override bool Equals(object obj) => Equals(obj as ExtendedAddress);
+        public override int GetHashCode() {
+            var hashCode = new HashCode();
+            hashCode.Add(Street);
+            hashCode.Add(StreetExtra);
+            hashCode.Add(PostalCode);
+            hashCode.Add(City);
+            hashCode.Add(Country);
+            hashCode.Add((int) AddressType);
+            hashCode.Add(Flags);
+            hashCode.Add(AddressType2);
+            return hashCode.ToHashCode();
+        }
     }
 }