+++ demo/DemoV2/Model/GroupedOneOf.generated.cs
@@ -11,7 +11,7 @@ using DemoV2.Model.Base;
 
 namespace DemoV2.Model {
 /// AUTOGENERED BY caffoa ///
-    public partial class GroupedOneOf {
+    public partial class GroupedOneOf : IEquatable<GroupedOneOf> {
         public const string GroupedOneOfObjectName = "groupedOneOf";
         [JsonProperty("element")]
         public virtual AnyUser Element { get; set; }
@@ -25,5 +25,16 @@ namespace DemoV2.Model {
             AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, object>(other.AdditionalProperties) : null;
         }
         public GroupedOneOf ToGroupedOneOf() => new GroupedOneOf(this);
+        public bool Equals(GroupedOneOf other) {
+            if (ReferenceEquals(null, other)) return false;
+            if (ReferenceEquals(this, other)) return true;
+            return Element.Equals(other.Element);
+        }
+        public override bool Equals(object obj) => Equals(obj as GroupedOneOf);
+        public override int GetHashCode() {
+            var hashCode = new HashCode();
+            hashCode.Add(Element);
+            return hashCode.ToHashCode();
+        }
     }
 }