+++ demo/DemoV2/Model/TagInfos.generated.cs
@@ -11,7 +11,7 @@ using DemoV2.Model.Base;
 
 namespace DemoV2.Model {
 /// AUTOGENERED BY caffoa ///
-    public partial class TagInfos {
+    public partial class TagInfos : IEquatable<TagInfos> {
         public const string TagInfosObjectName = "tagInfos";
         [JsonProperty("user")]
         public virtual Dictionary<string, List<Guid>> User { get; set; } = new Dictionary<string, List<Guid>>();
@@ -25,5 +25,16 @@ namespace DemoV2.Model {
             AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, object>(other.AdditionalProperties) : null;
         }
         public TagInfos ToTagInfos() => new TagInfos(this);
+        public bool Equals(TagInfos other) {
+            if (ReferenceEquals(null, other)) return false;
+            if (ReferenceEquals(this, other)) return true;
+            return User.SequenceEqual(other.User);
+        }
+        public override bool Equals(object obj) => Equals(obj as TagInfos);
+        public override int GetHashCode() {
+            var hashCode = new HashCode();
+            hashCode.Add(User);
+            return hashCode.ToHashCode();
+        }
     }
 }